Update CI metric comparison to compare with target branch instead of `main`
