+++ .gitignore
@@ -38,3 +38,6 @@ htmlcov*/
 prof/
 sandbox/
 /designs
+node_modules/
+yarn.lock
+package.json
